requirements.txt
numpy>=1.24
requests>=2.32
pypdf>=4.2
fastapi>=0.112
uvicorn>=0.30
python-dotenv>=1.0
